requirements.txt
# Data Pipeline & Core Dependencies
requests>=2.28.0
pandas>=1.5.0
pyarrow>=12.0.0
fastparquet>=2023.1.0
python-dotenv>=0.21.0

# Scientific Computing
numpy>=1.23.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Machine Learning (Optional for LSTM model)
tensorflow>=2.12.0
keras>=2.12.0
